Fix unasync misses
