+++ bindings/python/pymongocrypt/asynchronous/explicit_encrypter.py
@@ -12,117 +12,9 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from pymongocrypt.asynchronous.mongocrypt import MongoCrypt
+from pymongocrypt.asynchronous.mongocrypt import AsyncMongoCrypt
 from pymongocrypt.asynchronous.state_machine import run_state_machine
-
-
-class ExplicitEncryptOpts(object):
-    def __init__(self, algorithm, key_id=None, key_alt_name=None,
-                 query_type=None, contention_factor=None, range_opts=None,
-                 is_expression=False):
-        """Options for explicit encryption.
-
-        :Parameters:
-          - `algorithm` (str): The algorithm to use.
-          - `key_id`: The data key _id.
-          - `key_alt_name` (bytes): Identifies a key vault document by
-            'keyAltName'. Must be BSON encoded document in the form:
-            { "keyAltName" : (BSON UTF8 value) }
-          - `query_type` (str): The query type to execute.
-          - `contention_factor` (int): The contention factor to use
-            when the algorithm is "Indexed".
-          - `range_opts` (bytes): Options for explicit encryption
-            with the "rangePreview" algorithm encoded as a BSON document.
-          - `is_expression` (boolean): True if this is an encryptExpression()
-            context. Defaults to False.
-
-        .. versionchanged:: 1.3
-           Added the `query_type` and `contention_factor` parameters.
-        .. versionchanged:: 1.5
-           Added the `range_opts` and `is_expression` parameters.
-        """
-        self.algorithm = algorithm
-        self.key_id = key_id
-        self.key_alt_name = key_alt_name
-        if query_type is not None:
-            if not isinstance(query_type, str):
-                raise TypeError(
-                    'query_type must be str or None, not: %r' % (type(query_type),))
-        self.query_type = query_type
-        if contention_factor is not None and not isinstance(contention_factor, int):
-            raise TypeError(
-                'contention_factor must be an int or None, not: %r' % (type(contention_factor),))
-        self.contention_factor = contention_factor
-        if range_opts is not None and not isinstance(range_opts, bytes):
-            raise TypeError(
-                'range_opts must be an bytes or None, not: %r' % (type(range_opts),))
-        self.range_opts = range_opts
-        self.is_expression = is_expression
-
-
-class DataKeyOpts(object):
-    def __init__(self, master_key=None, key_alt_names=None, key_material=None):
-        """Options for creating encryption keys.
-
-        :Parameters:
-          - `master_key`: Identifies a KMS-specific key used to encrypt the
-            new data key. If the kmsProvider is "local" the `master_key` is
-            not applicable and may be omitted.
-
-            If the `kms_provider` is "aws" it is required and has the
-            following fields::
-
-              - `region` (string): Required. The AWS region, e.g. "us-east-1".
-              - `key` (string): Required. The Amazon Resource Name (ARN) to
-                 the AWS customer.
-              - `endpoint` (string): Optional. An alternate host to send KMS
-                requests to. May include port number, e.g.
-                "kms.us-east-1.amazonaws.com:443".
-
-            If the `kms_provider` is "azure" it is required and has the
-            following fields::
-
-              - `keyVaultEndpoint` (string): Required. Host with optional
-                 port, e.g. "example.vault.azure.net".
-              - `keyName` (string): Required. Key name in the key vault.
-              - `keyVersion` (string): Optional. Version of the key to use.
-
-            If the `kms_provider` is "gcp" it is required and has the
-            following fields::
-
-              - `projectId` (string): Required. The Google cloud project ID.
-              - `location` (string): Required. The GCP location, e.g. "us-east1".
-              - `keyRing` (string): Required. Name of the key ring that contains
-                the key to use.
-              - `keyName` (string): Required. Name of the key to use.
-              - `keyVersion` (string): Optional. Version of the key to use.
-              - `endpoint` (string): Optional. Host with optional port.
-                Defaults to "cloudkms.googleapis.com".
-
-            If the `kms_provider` is "kmip" it is optional and has the
-            following fields::
-
-              - `keyId` (string): Optional. `keyId` is the KMIP Unique
-                Identifier to a 96 byte KMIP Secret Data managed object. If
-                keyId is omitted, the driver creates a random 96 byte KMIP
-                Secret Data managed object.
-              - `endpoint` (string): Optional. Host with optional
-                 port, e.g. "example.vault.azure.net:".
-
-          - `key_alt_names`: An optional list of bytes suitable to be passed to
-            mongocrypt_ctx_setopt_key_alt_name. Each element must be BSON
-            encoded document in the form: { "keyAltName" : (BSON UTF8 value) }
-
-          - `key_material`: An optional binary value of 96 bytes to use as
-            custom key material for the data key being created. If
-            ``key_material`` is given, the custom key material is used for
-            encrypting and decrypting data. Otherwise, the key material for the
-            new data key is generated from a cryptographically secure random
-            device.
-        """
-        self.master_key = master_key
-        self.key_alt_names = key_alt_names
-        self.key_material = key_material
+from pymongocrypt.options import DataKeyOpts, ExplicitEncryptOpts
 
 
 class AsyncExplicitEncrypter(object):
@@ -139,7 +31,7 @@ class AsyncExplicitEncrypter(object):
         self.callback = callback
         if mongo_crypt_opts.schema_map is not None:
             raise ValueError("mongo_crypt_opts.schema_map must be None")
-        self.mongocrypt = MongoCrypt(mongo_crypt_opts, callback)
+        self.mongocrypt = AsyncMongoCrypt(mongo_crypt_opts, callback)
 
     async def create_data_key(self, kms_provider, master_key=None,
                         key_alt_names=None, key_material=None):
