+++ bindings/python/pyproject.toml
@@ -43,7 +43,7 @@ dependencies = [
     "cffi>=1.12.0,<2",
     "cryptography>=40",
     "packaging>=21.0",
-    "httpx<0.25.0",
+    "httpx",
 ]
 
 [project.urls]
